requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 8.2.0

albumentations == 1.0.0

matplotlib == 3.3.4

numpy == 1.20.1

opencv_python_headless == 4.5.2.54

pandas == 1.2.4

scikit_image == 0.18.1

scikit_learn == 0.24.1

seaborn == 0.11.1

streamlit == 0.84.0

timm == 0.4.9

torch == 1.9.0

torchvision == 0.10.0

tqdm == 4.59.0
